==> src/Mayer/04_spatial_models.Rmd <==
---
  title: "Spatial Models, Mayer"
author: "Claire Kelling"
date: "June 11, 2018"
output: html_document
editor_options: 
  chunk_output_type: console
---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

The purpose of this code is to fit the CAR model to voting data, aggregated by ward. We only have Census data by district and/or block group so this presents a spatial mismatch file. 

First, we will install all necessary packages.
```{r}
# Packages:
library(sp)
library(spdep)
library(classInt)
library(fields)
library(ggplot2)
library(dplyr)
library(ade4) 
library(igraph) 
library(CARBayesdata)
library(CARBayes)
library(gridExtra)
library(xtable)
library(stringi)
require(rgdal)
library(ngspatial)
library(plyr)
library(readxl)

```

Now, we will load the data, including the following:
  * shape files
* demographics
* voting data
```{r}
#ward shape files
#wi_ward
# Location of ward group data:
# C:\Users\ckell\Dropbox\Gerrycamp Regressions Project\Mayer Project\Data to be Used in R\CombinedWIData.xslx


ward_shp <- readOGR(dsn = "C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/data", layer = "wiward")

#voting and demographic data
#quick fit with demographics
#demog_vote <- read.csv("C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/data/QuickFitWithDemographics.csv")

ForR <- read_excel("C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/Data to be Used in R/ForR.xlsx")

```

Now we need to join the spatial polygons and the demographic/voting data.
```{r}
#re-formatting to add the data to the 
ward_shp@data <- ward_shp@data[,-c(14:ncol(ward_shp@data))]

#check classes
class(ward_shp@data$OBJECTID_1)

#change column name for ID variables for which we will join
#colnames(demog_vote)[2] <- "GEOID"
colnames(ward_shp@data)[1] <- "GEOID"
colnames(ForR)[5] <- "GEOID"

#change to character variable
#demog_vote$GEOID <- as.character(demog_vote$GEOID)
ward_shp@data$GEOID <- as.character(ward_shp@data$GEOID)
ForR$GEOID <- as.character(ForR$GEOID)

#creating new variable
#demog_vote$IncumD_Pop <- demog_vote$Dincumbent*demog_vote$PERSONS18
#demog_vote$IncumR_Pop <- demog_vote$Rincumbent*demog_vote$PERSONS18

#join on GEOID
ward_shp@data <- left_join(ward_shp@data, ForR, by = (GEOID = "GEOID"))

#BBPRESDEM12 + BBPRESREP12+PERSONS18 +BLACK18 + WHITE18 + HISPANIC18+ IncumD_Pop + IncumR_Pop

# class(ward_shp@data$BBWSADEM12) 
# class(ward_shp@data$BBWSADEM12) 
# class(ward_shp@data$BBPRESDEM12) 
# class(ward_shp@data$BBPRESREP12)


# ward_shp@data$BBWSADEM12 <- round(as.numeric(ward_shp@data$BBWSADEM12))
# ward_shp@data$BBWSADEM12 <- round(as.numeric(ward_shp@data$BBWSADEM12))
# ward_shp@data$BBPRESDEM12 <- round(as.numeric(ward_shp@data$BBPRESDEM12))
# ward_shp@data$BBPRESREP12 <- round(as.numeric(ward_shp@data$BBPRESREP12))
# ward_shp@data$CWSADEM12 <- round(ward_shp@data$CWSADEM12)
# ward_shp@data$CWSAREP12 <- round(ward_shp@data$CWSAREP12)

ward_shp@data$c_wsa_dem12 <- round(ward_shp@data$c_wsa_dem12)
ward_shp@data$c_wsa_rep12 <- round(ward_shp@data$c_wsa_rep12)

#we have successfully joined the two datasets
```


Our first step in terms of modeling will be to run our model in a linear model and to test for spatial dependence. 
```{r}

#need to delete Object ID 6481 (outlier)
#range(ward_shp@data$c_wsa_dem12)
#out <- which(ward_shp@data$GEOID == 6481)
#ward_shp <- ward_shp[-out,]
range(ward_shp@data$c_wsa_dem12)
range(ward_shp@data$c_wsa_rep12)

#we need to decide variables to use in our analysis
#colnames(demog_vote)

#create a variable for uncontested races
#unc <- unique(which(ward_shp@data$Contested == "U"))
unc <- unique(which(ward_shp@data$contested_1 == "0"))
con_shp <- ward_shp[-unc,]

#non-spatial modeling
#d.form <- CWSADEM12 ~ CPREDEM12 + CPREREP12+PERSONS18 +BLACK18  + HISPANIC18+ IncumD_Pop + IncumR_Pop 
#r.form <- CWSAREP12 ~ CPREDEM12 + CPREREP12+PERSONS18 +BLACK18  + HISPANIC18+ IncumD_Pop + IncumR_Pop 

#d.form <- WSADEM12 ~ PREDEM12 + PREREP12+total_vep +black_vep + hisp_vep + dem_incum + rep_incum 
#r.form <- WSAREP12 ~ PREDEM12 + PREREP12+total_vep +black_vep + hisp_vep + dem_incum + rep_incum 

#trying so that there is no overfit
d.form <- WSADEM12 ~ PREREP12  +black_vep + hisp_vep + dem_incum + rep_incum 
r.form <- WSAREP12 ~ PREDEM12 + total_vep +black_vep + hisp_vep + dem_incum + rep_incum 

#create linear models
c.d.model <- lm(formula=d.form, data=con_shp@data)
c.d.resid.model <- con_shp@data$c_wsa_dem12 - c.d.model$fitted.values
c.r.model <- lm(formula=r.form, data=con_shp@data)
c.r.resid.model <- con_shp@data$c_wsa_rep12 - c.r.model$fitted.values

#test for spatial dependence
#null hypothesis of no spatial autocorrelation (alternative of positive spatial autocorrelation)
#also computes Moran's I statistic 
#if p-value < 0.05, we conclude there is positve spatial autocorrelation
W.nb.con <- poly2nb(con_shp, row.names = rownames(con_shp@data))

##############
### Creating adjacency matrix from nb
##############
W_mat_con <- nb2mat(W.nb.con, style="B", zero.policy=TRUE)


#############
### Moran's I Test
#############

W.list.con <- nb2listw(W.nb.con, style="B", zero.policy = TRUE)

moran.mc(x=c.d.resid.glm, listw=W.list.con, nsim=1000, zero.policy = TRUE)
moran.mc(x=c.r.resid.glm, listw=W.list.con, nsim=1000, zero.policy = TRUE)

#if p-value < 0.05, we conclude there is positve spatial autocorrelation

```



Need to eliminate the islands with no neighbors
```{r}
no_nb <- which(rowSums(W_mat_con) == 0) #13 with no neighbors

nonb_W <- W_mat_con[-no_nb, -no_nb]
nonb_shp <- con_shp[-no_nb,]

nonb_Wnb <- poly2nb(nonb_shp, row.names = rownames(nonb_shp@data))
```


Now, we will fit both the Leroux and the BYM model.
```{r}
##############
#Leroux Model
##############
#started at 3:07pm, finished in  2162.1 seconds
c.d.model.ler <- S.CARleroux(formula=d.form, data=nonb_shp@data,
                             family="poisson", W=nonb_W, burnin=20000, n.sample=120000, thin=10)
#started at 3:50pm, started MCMC at 3:58pm, finished in  2595 seconds
c.r.model.ler <- S.CARleroux(formula=r.form, data=nonb_shp@data,
                             family="poisson", W=nonb_W, burnin=20000, n.sample=120000, thin=10)
#need to rerun the d leroux model

##############
# BYM Model
##############
#started at 4:48, 47% at 5:15, finished in  3378.8 seconds
c.d.model.bym <- S.CARbym(formula=d.form, data=nonb_shp@data,
                             family="poisson", W=nonb_W, burnin=20000, n.sample=120000, thin=10)

#started at 5:45pm, finished in  3422 seconds
c.r.model.bym <- S.CARbym(formula=r.form, data=nonb_shp@data,
                             family="poisson", W=nonb_W, burnin=20000, n.sample=120000, thin=10)


#save(c.d.model.ler, #c.r.model.ler, c.d.model.bym, c.r.model.bym, 
#     file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_d_model_ler.Rdata")
#save(c.r.model.ler, #c.r.model.ler, c.d.model.bym, c.r.model.bym, 
#     file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_r_model_ler.Rdata")
#save(c.d.model.bym, #c.r.model.ler, c.d.model.bym, c.r.model.bym, 
#     file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_d_model_bym.Rdata")
#save(c.r.model.bym, #c.r.model.ler, c.d.model.bym, c.r.model.bym, 
#     file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_r_model_bym.Rdata")

#rm(c.d.model.ler)
#rm(c.r.model.ler)
#rm(c.d.model.bym)
#rm(c.r.model.bym)

#     load(file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_d_model_ler.Rdata")
#     load(file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_r_model_ler.Rdata")
#     load(file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_d_model_bym.Rdata")
#     load(file = "C:/Users/ckell/Desktop/Summer 2018/MGGG/Regression/data2/c_r_model_bym.Rdata")

##############
# Model Evaluation
##############
#Only using contested precincts:
#  dem ler
c.d.model.ler$modelfit
#  rep ler
c.r.model.ler$modelfit
#  dem bym
c.d.model.bym$modelfit
#  rep bym
c.r.model.bym$modelfit
```

Now, I will also include a model that accounts for spatial autocorrelation.
```{r}
#d.sp.sglmm.fit <- sparse.sglmm(formula = d.form,data=no_dem@data, family = poisson, A = W_mat_nod,
#                             verbose = TRUE) #tune = list(sigma.s = 0.02)

#r.sp.sglmm.fit <- sparse.sglmm(formula = r.form,data=no_rep@data, family = poisson, A = W_mat_nor,
#                             verbose = TRUE) #tune = list(sigma.s = 0.02)

#rownames(nonb_W) <- NULL
#c.d.sp.sglmm.fit <- sparse.sglmm(formula = d.form,data=nonb_shp@data, family = poisson, 
#                                 A = nonb_W, verbose = TRUE) #tune = list(sigma.s = 0.02)

#25 min at 9:43, stuck at 44%, 23m 49s

#c.r.sp.sglmm.fit <- sparse.sglmm(formula = r.form,data=nonb_shp@data, family = poisson, 
#                                 A = nonb_W, verbose = TRUE) #tune = list(sigma.s = 0.02)
#11%, 32m 22s remaining at 9:56pm
#55% ~19m 27s  

#save the models for later reference, takes over an hour to run
#save(c.d.model.ler, c.r.model.ler, c.d.model.bym, c.r.model.bym, 
#     file = "C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/prelim_fit/prelim_mod_fit2.Rdata")

c.d.sp.sglmm.fit$dic
c.d.sp.sglmm.fit$pD

c.r.sp.sglmm.fit$dic
c.r.sp.sglmm.fit$pD

```

Now, we need to compare coefficients to see if there was any spatial autocorrelation.
```{r}
model.bym$summary.results[,1:3]

summary(c.d.sp.sglmm.fit)
```

#Model Evaluation
Now, I will compare the models using mean-squared prediction error.

First, I need to set up predicted for each model and actual for both Democratic and Republican races.
```{r}
actual.d <- nonb_shp@data$WSADEM12
actual.r <- nonb_shp@data$WSAREP12

d.pred.bym <- c.d.model.bym$fitted.values
r.pred.bym <- c.r.model.bym$fitted.values

d.pred.ler <- c.d.model.ler$fitted.values
r.pred.ler <- c.r.model.ler$fitted.values

#d.pred.sglmm <- c.d.sp.sglmm.fit$fitted.values
#r.pred.sglmm <- c.r.sp.sglmm.fit$fitted.values

#78 districts  (uncontested)
con_shp@data$ASM <- as.numeric(con_shp@data$ASM)
actual_tab <- cbind(con_shp@data$ASM, con_shp@data$WSADEM12, con_shp@data$WSAREP12)
actual_tab <- as.data.frame(actual_tab)

colnames(actual_tab) <- c("District", "actual_d", "actual_r")
actual_tab2 <- actual_tab %>% group_by(District) %>% dplyr::summarise(dem_votes=sum(actual_d), rep_votes = sum(actual_r))

length(which(actual_tab2[,2] > actual_tab2[,3])) #16 times that the dem was greater

new_var <- rep(NA, length(actual.r))
new_var[which(actual.r- r.pred.bym > 5)] <- 1
sum(new_var, na.rm=T)
# 325
View(cbind(r.pred.bym, actual.r, nonb_shp@data$dem_incum, nonb_shp@data$rep_incum, new_var))
View(cbind(r.pred.bym, actual.r, nonb_shp@data$PREDEM12, nonb_shp@data$PREREP12, new_var))
View(cbind(r.pred.bym, actual.r, nonb_shp@data$total_vep, nonb_shp@data$black_vep, new_var))

View(cbind(d.pred.ler, actual.d))

View(cbind(r.pred.ler, actual.r))

range(r.pred.ler - actual.r)

View(cbind(r.pred.ler, actual.r, r.pred.ler - actual.r, nonb_shp@data$total_vep))

percentile <- ecdf(nonb_shp@data$persons_18)
test <- percentile(nonb_shp@data$persons_18)

#should compare percentiles to the error


```

Now, I will compute the MSPE for each model.
```{r}
#### Democratic Prediction:
#bym
d.bym.mspe <- sqrt(mean((d.pred.bym-actual.d)^2))
#leroux
d.ler.mspe <- sqrt(mean((d.pred.ler-actual.d)^2))
#sglmm
#d.sglmm.mspe <- mean((d.pred.sglmm - actual.d)^2)

#### Republican Prediction:
#bym
r.bym.mspe <- sqrt(mean((r.pred.bym-actual.r)^2))
#leroux
r.ler.mspe <- sqrt(mean((r.pred.ler-actual.r)^2))
#sglmm
#r.sglmm.mspe <- mean((r.pred.sglmm - actual.r)^2)

mspe_tab <- rbind(c(d.bym.mspe, d.ler.mspe),#, d.sglmm.mspe),
                  c(r.bym.mspe, r.ler.mspe))#, r.sglmm.mspe))
#mspe_tab_count <- mspe_tab
colnames(mspe_tab) <- c("bym", "leroux")#, "sglmm")
rownames(mspe_tab) <- c("Democrat", "Republican")

xtable(mspe_tab)

```


Bayesian Model Comparison Techniques:
```{r}
#smaller DIC is better
#higher perc dev expl is better
#leroux models are much better and explain much of the deviance (than bym)
# dem ler
c.d.model.ler$modelfit
#  rep ler
c.r.model.ler$modelfit
#  dem bym
c.d.model.bym$modelfit
#  rep bym
c.r.model.bym$modelfit

bayes_mod_eval <- rbind(c(c.d.model.bym$modelfit[1], c.d.model.ler$modelfit[1], 
                          c.r.model.bym$modelfit[1], c.r.model.ler$modelfit[1]),
                        c(c.d.model.bym$modelfit[7], c.d.model.ler$modelfit[7], 
                          c.r.model.bym$modelfit[7], c.r.model.ler$modelfit[7]))
bayes_mod_eval <- as.data.frame(bayes_mod_eval)
colnames(bayes_mod_eval) <- c("BYM Dem", "Ler Dem", "BYM Rep", "Ler Rep")
rownames(bayes_mod_eval) <- c("DIC", "perc dev explained")

#c.d.sp.sglmm.fit$dic
#c.d.sp.sglmm.fit$pD

#c.r.sp.sglmm.fit$dic
#c.r.sp.sglmm.fit$pD
```


Creation of counterfactual scenario where there is no incumbent:
```{r}
#First, I will create the model dataframe
#c_wsa_rep12 ~ c_pre_dem12 + c_pre_rep12+total_vep +black_vep + hisp_vep + dem_incum + rep_incum 

mod_dat <- demog_vote[,c("c_pre_dem12", "c_pre_rep12", "total_vep", "black_vep", "hisp_vep", "dem_incum", "rep_incum")]
mod_dat$int <- rep(1, nrow(mod_dat))
mod_dat <- cbind(mod_dat[,9], mod_dat[,1:8])

mod_dat2 <- con_shp@data[,c("c_pre_dem12", "c_pre_rep12", "total_vep", "black_vep", "hisp_vep", "dem_incum", "rep_incum")]
mod_dat2$int <- rep(1, nrow(mod_dat2))
mod_dat2 <- cbind(mod_dat2[,9], mod_dat2[,1:8])

#create function that returns predicted values under counterfactual situation
pred_val <- function(coeffs.Dem, coeffs.Rep, mod_dat, expon){
  #mod_dat <- mod_dat2
  coeffs.Dem[8] <- 0
  coeffs.Dem[9] <- 0
  coeffs.Rep[8] <- 0
  coeffs.Rep[9] <- 0
  
  #test2 <- data.frame(lapply(test, function(x) as.numeric(as.character(x))))
  tableDem <- as.matrix(mod_dat)%*%coeffs.Dem
  tableDem <- as.numeric(tableDem)
  if(expon = T){
    tableDem <- exp(tableDem) #this is when we have a GLM framework
  }
  tableDem <- cbind(tableDem, demog_vote$ASM)
  tableDem <- as.data.frame(tableDem)
  colnames(tableDem) <- c("votes", "District")
  
  length(unique(tableDem$District))
  
  v_table_d <- tableDem %>% group_by(District) %>% dplyr::summarise(dem_votes=sum(votes))
  
  tableRep <- as.matrix(mod_dat)%*%coeffs.Rep
  tableRep <- as.numeric(tableRep)
  if(expon = T){
    tableRep <- exp(tableRep) #this is when we have a GLM framework 
  }
  tableRep <- cbind(tableRep, demog_vote$ASM)
  tableRep <- as.data.frame(tableRep)
  colnames(tableRep) <- c("votes", "District")
  
  v_table_r <- tableRep %>% group_by(District) %>% dplyr::summarise(rep_votes=sum(votes))
  
  v_table <- left_join(v_table_d, v_table_r)
  
  colnames(v_table) <- c("District", "Dem. Votes", "Rep. Votes")
  v_table2 <- v_table
  return(v_table)
}

######
#BYM:
######

coeffs.Dem <- c.d.model.bym$summary.results[1:9,1]
coeffs.Rep <- c.r.model.bym$summary.results[1:9,1]

coeffs.Dem <- as.numeric(coeffs.Dem)
coeffs.Rep <- as.numeric(coeffs.Rep)

v_table_bym <- pred_val(coeffs.Dem, coeffs.Rep, mod_dat, T)
length(which(v_table_bym[,2] > v_table_bym[,3])) #still only predicts 3

######
#Leroux:
######

coeffs.Dem <- c.d.model.ler$summary.results[1:9,1]
coeffs.Rep <- c.r.model.ler$summary.results[1:9,1]

coeffs.Dem <- as.numeric(coeffs.Dem)
coeffs.Rep <- as.numeric(coeffs.Rep)

v_table_ler <- pred_val(coeffs.Dem, coeffs.Rep, mod_dat, T)
length(which(v_table_ler[,2] > v_table_ler[,3])) #still only predicts 2

######
#SGLMM:
######

# coeffs.Dem <- c.d.sp.sglmm.fit$coefficients
# coeffs.Rep <- c.r.sp.sglmm.fit$coefficients
# 
# coeffs.Dem <- as.numeric(coeffs.Dem)
# coeffs.Rep <- as.numeric(coeffs.Rep)
# 
# v_table_sglmm <- pred_val(coeffs.Dem, coeffs.Rep)

#write.csv(v_table_lm, file = 'C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/data/mod_fit/lin_mod.csv')
#write.csv(v_table_glm, file = 'C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/data/mod_fit/glm.csv')
#write.csv(v_table_bym, file = 'C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/data/mod_fit/bym.csv')
#write.csv(v_table_ler, file = 'C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/data/mod_fit/ler.csv')
#write.csv(v_table_sglmm, file = 'C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/data/mod_fit/sglmm.csv')```
```


Now, I will print model output for our presentation.
```{r}
#Bayesian model comparison
xtable(bayes_mod_eval)

#mspe
xtable(mspe_tab)

```


```{r}
plot(nonb_shp)

## Plot for Proportion
#prec_shp@data   <- join(prec_shp@data, data, by="CD_GEOCODI")
nonb_shp@data$error <- abs(r.pred.ler- actual.r)
nonb_shp@data$id <- rownames(nonb_shp@data)
sp_f     <- fortify(nonb_shp)
sp_f     <- join(sp_f,nonb_shp@data, by="id")
#sp_f <- sp_f[which(sp_f$long < 1e10),]

fit_by_ward <- ggplot()+ 
  #geom_polygon(data=dist_shp,aes(long,lat, group = group), fill = NA, col = "red") + 
  geom_polygon(data = sp_f, aes(long, lat, group = group, fill = log(error))) + 
  coord_equal() +
  labs(fill = "log error by Ward") +
  ggtitle("absolute Error by Ward")+ scale_fill_gradient2(midpoint = 0.5,low = "white", high = "red")


nonb_shp@data$d_error <- abs(d.pred.bym- actual.d)
nonb_shp@data$id <- rownames(nonb_shp@data)
sp_f     <- fortify(nonb_shp)
sp_f     <- join(sp_f,nonb_shp@data, by="id")
#sp_f <- sp_f[which(sp_f$long < 1e10),]

fit_by_ward <- ggplot()+ 
  #geom_polygon(data=dist_shp,aes(long,lat, group = group), fill = NA, col = "red") + 
  geom_polygon(data = sp_f, aes(long, lat, group = group, fill = log(d_error))) + 
  coord_equal() +
  labs(fill = "Dem log error by Ward") +
  ggtitle("Dem absolute Error by Ward")+ scale_fill_gradient2(midpoint = 0.5,low = "white", high = "red")

model_error <- nonb_shp@data
write.csv(model_error, file = "C:/Users/ckell/Dropbox/Gerrycamp Regressions Project/Mayer Project/model_error.csv")

r.form
#WSAREP12 ~ PREDEM12 + PREREP12 + total_vep + black_vep + hisp_vep + 
#    dem_incum + rep_incum

plot(model_error$error, model_error$total_vep)
plot(model_error$error, model_error$black_vep)
plot(model_error$error, model_error$hisp_vep)

plot(model_error$error, model_error$dem_incum)
plot(model_error$error, model_error$rep_incum)

plot(model_error$error, model_error$PREDEM12)
plot(model_error$error, model_error$PREREP12)
plot(model_error$error, model_error$WSADEM12)
plot(model_error$error, model_error$WSAREP12)


```


Check for overfit? Try to fit models without PREDEM12 or PREREP12


Need to check the non counterfactual scenario to check efficiency gap

check difference in estimate for linear model
